Add sanctum middleware

+++ routes/api.php
@@ -1,6 +1,5 @@
 <?php
 
-use App\Http\Resources\Book\BookResource;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -20,15 +19,21 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 });
 
 Route::group(['namespace' => 'App\Http\Controllers\Api'], function () {
+    Route::get('/check_user', function (Request $request) {
+        $name = $request->user()->tokens;
+        return $name;
+    })->middleware(['auth:sanctum']);
+
     Route::group(['namespace' => 'Book'], function () {
         Route::get('/book', 'BookController@books');
         Route::get('/book/{book}', 'BookController@show');
-        Route::patch('/book/{book}', 'BookController@update');
-        Route::delete('/book/{book}', 'BookController@delete');
+        Route::patch('/book/{book}', 'BookController@update')->middleware(['auth:sanctum']);
+        Route::delete('/book/{book}', 'BookController@delete')->middleware(['auth:sanctum']);
     });
     Route::group(['namespace' => 'User'], function () {
+        Route::post('/login', 'LoginController');
         Route::get('/user', 'UserController@users');
         Route::get('/user/{user}', 'UserController@show');
-        Route::patch('/user/{user}', 'UserController@update');
+        Route::patch('/user/{user}', 'UserController@update')->middleware(['auth:sanctum']);
     });
 });
